lambda/postgis-extension/requirements.txt
aws-psycopg2==1.3.8
boto3>=1.26.0
